Extracting responsive logic to a custom hook
